fix: resolve build errors in workflow execution code
- Fix ActionInfo import (model package, not spi)
- Move workflow instance endpoints from separate resource to
  ProjectsResourceImpl (matches generated interface)
- Add additionalProperties to OpenAPI freeform object schemas
  so generated beans support dynamic JSON properties
- Fix null values in workflow context map (Map.copyOf rejects nulls)
- Fix test expecting 201 from project creation (actual is 200)
